Refactor Instructor Tasks - in prep for Grade report work
