+++ lms/djangoapps/grades/context.py
@@ -9,7 +9,7 @@ from .scores import possibly_scored
 
 def grading_context_for_course(course_key):
     """
-    Same as grading_context, but takes in a course object.
+    Same as grading_context, but takes in a course key.
     """
     course_structure = get_course_in_cache(course_key)
     return grading_context(course_structure)
